+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Add
+ - Add `Dockerfile` for cmake example
+ - Add `all_codecs` feature
+
 ## [0.8.0] - 2024-08-05
 
 ### Add