+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Add `Dockerfile` for cmake example
  - Add `all_codecs` feature
 
+### Changed 
+ - Fetch deps at configure time in `Findzarrs.cmake`
+
 ## [0.8.0] - 2024-08-05
 
 ### Add