+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.1] - 2024-08-05
+
 ### Add
  - Add `Dockerfile` for cmake example
  - Add `all_codecs` feature
@@ -126,7 +128,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Initial public release
 
-[unreleased]: https://github.com/LDeakin/zarrs_ffi/compare/v0.7.2...HEAD
+[unreleased]: https://github.com/LDeakin/zarrs_ffi/compare/v0.8.1...HEAD
+[0.8.1]: https://github.com/LDeakin/zarrs_ffi/releases/tag/v0.8.1
+[0.8.0]: https://github.com/LDeakin/zarrs_ffi/releases/tag/v0.8.0
 [0.7.2]: https://github.com/LDeakin/zarrs_ffi/releases/tag/v0.7.2
 [0.7.1]: https://github.com/LDeakin/zarrs_ffi/releases/tag/v0.7.1
 [0.7.0]: https://github.com/LDeakin/zarrs_ffi/releases/tag/v0.7.0